requirements.txt
altgraph==0.17.3
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
djangorestframework==3.13.1
future==0.18.2
idna==3.3
jsonfield==3.1.0
oauthlib==3.2.0
pefile==2022.5.30
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
swapper==1.3.0
tzdata==2022.1
urllib3==1.26.11
